Multiple changes
 - Created resource file to save exception messages
 - Changed http types
 - Added more try catches
 - Changed request url in react
 - Changed url endpoint in api controller

+++ TicketBriteAPI/TicketBrite.Data/Repositories/ShoppingCartRepository.cs
@@ -20,6 +20,11 @@ namespace TicketBrite.Data.Repositories
         {
             ReservedTicket ticket = _context.ReservedTickets.Find(reservationID);
 
+            if(ticket == null)
+            {
+                throw new KeyNotFoundException("Gereserveerde ticket niet gevonden!");
+            }
+
             _context.ReservedTickets.Remove(ticket);
             _context.SaveChanges();
         }
